wrapped domain logic in classes and interfaces
