+++ upload_controller_service/domain/__init__.py
@@ -17,15 +17,3 @@
 This sub-package contains the main business functionality of this service.
 It should not contain any service API-related code.
 """
-
-from .main import (  # noqa: F401
-    FileAlreadyInInboxError,
-    FileAlreadyRegisteredError,
-    FileNotInInboxError,
-    FileNotReadyForConfirmUpload,
-    FileNotRegisteredError,
-    confirm_file_upload,
-    get_upload_url,
-    handle_file_registered,
-    handle_new_study,
-)
